cfl kernel
